+++ kernel/headers/CPUKernel.h
@@ -2,6 +2,6 @@
 #define CPUKERNEL_H_
 #include "kernel.h"
 
-uint16_t get_pid_from_cpu(int fd, op_code instruccion);
+int get_pid_from_cpu(int fd, op_code instruccion);
 
 #endif /* CPUKERNEL_H_ */
