+++ utils/headers/utils.h
@@ -15,7 +15,7 @@
 
 bool config_has_all_properties(t_config*, char**);
 
-bool send_un_char_y_un_int(int fd, char* cadena, uint16_t entero);
+bool send_un_char_y_un_int(int fd, char* cadena, int entero);
 bool send_dos_ints(int fd, uint8_t int1, uint8_t int2);
 bool send_string(int fd, char* string);
 bool send_data(int fd, void* data, size_t size);
